% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_patch_df.R
\name{create_patch_df}
\alias{create_patch_df}
\title{Create a dataframe that counts each patch independently for the prioritizr
input}
\usage{
create_patch_df(
  spatial_grid,
  features,
  patches,
  costs = NULL,
  locked_out = NULL,
  locked_in = NULL
)
}
\arguments{
\item{spatial_grid}{a raster or sf template with the desired resolution and
coordinate reference system; values in areas of interest are 1, while all
other values are NA}

\item{features}{a raster or sf object that includes all relevant features to
be used in the prioritization; each layer of the raster or each column of
the sf object identifies the location of each feature}

\item{patches}{a raster or sf object generated by \code{create_patches()}; a
single feature split into spatially distinct groups; each layer of the
raster or each column of the sf object identifies the location of each
patch}

\item{costs}{a raster or sf object of the costs for protecting each cell in
the planning grid}

\item{locked_out}{a raster or sf object for areas to be locked out
(absolutely not protected) in the prioritization}

\item{locked_in}{a raster or sf object for areas to be locked in (absolutely
protected) in the prioritization}
}
\value{
A data frame to be used as an input for \code{create_boundary_matrix()}
}
\description{
This function takes all of the inputs of prioritizr and creates
a dataframe precursor for prioritizr. This specific function includes the
use of "patches" which have been already generated by \code{create_patches()}.
"Patches" are spatial groups of a particular feature, such as seamounts.
Prioritizr will protect entire patches to meet the protection goal for that
feature. The output of this function should be passed to
\code{create_boundary_matrix()} before being passed to prioritizr.
}
\examples{
# Import some planning data

#import planning units/ cost raster
pu_raster <- terra::rast(system.file("extdata/pu_raster.tif", package = "patchwise"))
#import species distributions - feature data for planning
species_distributions <- terra::rast(system.file("extdata/spp_distributions.tif", package = "patchwise"))

#import seamounts
seamounts <- terra::rast(system.file("extdata/seamounts.tif", package = "patchwise"))

# Create seamount patches as multi-layer raster
patches_raster <- create_patches(seamounts)

patches_df <- create_patch_df(spatial_grid = pu_raster, features = species_distributions,
patches = patches_raster, costs = pu_raster)
}
