% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_targets.R
\name{features_targets}
\alias{features_targets}
\title{Create a dataframe for the relative and absolute targets for features}
\usage{
features_targets(
  targets,
  features,
  pre_patches,
  locked_out = NULL,
  locked_in = NULL
)
}
\arguments{
\item{targets}{a vector of targets for protection (range between 0 and 1);
must be the same length as the number of features + the pre-patches
variable}

\item{features}{a raster or sf object that includes all relevant features to
be used in the prioritization; each layer of the raster or each column of
the sf object identifies the location of each feature}

\item{pre_patches}{a raster or sf object that includes the feature that is to
be split into patches (not the layer that is already split into patches)}

\item{locked_out}{a raster or sf object for areas to be locked out
(absolutely not protected) in the prioritization}

\item{locked_in}{a raster or sf object for areas to be locked in (absolutely
protected) in the prioritization}
}
\value{
A data frame to be used to specify targets for features. Will need to
be plugged into \code{constraint_targets()} before being implemented into
prioritizr
}
\description{
This function creates a dataframe of targets for each of the
features to be used in the prioritization
}
\examples{
# Import some planning data

#import species distributions - feature data for planning
species_distributions <- terra::rast(system.file("extdata/spp_distributions.tif", package = "patchwise"))

#import seamounts
seamounts <- terra::rast(system.file("extdata/seamounts.tif", package = "patchwise"))


# Create targets for protection - use 20\% for each feature (including 20\% of whole seamounts) in this example
targets_df <- features_targets(targets = rep(0.2, (terra::nlyr(species_distributions) + 1)), features = species_distributions, pre_patches = seamounts)
head(targets_df)
}
