% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_boundary_matrix.R
\name{create_boundary_matrix}
\alias{create_boundary_matrix}
\title{Create boundary matrix for prioritizr if you are using patches}
\usage{
create_boundary_matrix(planning_grid, patches, patch_df)
}
\arguments{
\item{planning_grid}{a raster or sf template with the desired resolution and coordinate reference system generated by \code{offshoredatr::get_planning_grid()}; values in areas of interest are 1, while all other values are NA (only required if feature is a sf object)}

\item{patches}{a raster or sf object generated by \code{create_patches()}; a single feature split into spatially distinct groups; each layer of the raster or each column of the sf object identifies the location of each patch}

\item{patch_df}{a dataframe generated by \code{create_patch_df()} that includes constraints for each patch and grid cell combination}
}
\value{
A boundary matrix to plug into prioritizr
}
\description{
This function creates a boundary matrix to plug into prioritizr
}
\examples{
# Start with a little housekeeping to get the data from oceandatr
# Choose area of interest (Bermuda EEZ)
area <- oceandatr::get_area(area_name = "Bermuda")
projection <-'+proj=laea +lon_0=-64.8108333 +lat_0=32.3571917 +datum=WGS84 +units=m +no_defs'
# Create a planning grid
planning_raster <- oceandatr::get_planning_grid(area, projection = projection)
# Grab all relevant data
features_raster <- oceandatr::get_features(planning_grid = planning_raster)
# Separate seamount data - we want to protect entire patches
seamounts_raster <- features_raster[["seamounts"]]
features_raster <- features_raster[[names(features_raster)[names(features_raster) != "seamounts"]]]
# Create a "cost" to protecting a cell - just a uniform cost for this example
cost_raster <- setNames(planning_raster, "cost")
# Create patches from layer
patches_raster <- create_patches(seamounts_raster)
# Create patch dataframe
patches_raster_df <- create_patch_df(planning_grid = planning_raster, features = features_raster,
  patches = patches_raster, costs = cost_raster)
# Create boundary matrix for prioritizr
boundary_matrix <- create_boundary_matrix(planning_grid = planning_raster,
  patches = patches_raster, patch_df = patches_raster_df)
}
