% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_boundary_matrix.R
\name{create_boundary_matrix}
\alias{create_boundary_matrix}
\title{Create boundary matrix for prioritizr if you are using patches}
\usage{
create_boundary_matrix(planning_grid, patches, patch_df)
}
\arguments{
\item{planning_grid}{a raster or sf template with the desired resolution and coordinate reference system generated by \code{offshoredatr::get_planning_grid()}; values in areas of interest are 1, while all other values are NA (only required if feature is a sf object)}

\item{patches}{a raster or sf object generated by \code{create_patches()}; a single feature split into spatially distinct groups; each layer of the raster or each column of the sf object identifies the location of each patch}

\item{patch_df}{a dataframe generated by \code{create_patch_df()} that includes constraints for each patch and grid cell combination}
}
\value{
A boundary matrix to plug into prioritizr
}
\description{
This function creates a boundary matrix to plug into prioritizr
}
\examples{
## To be updated later
# Create patches from seamount raster data
seamount_patches <- create_patches(seamount_raster)
# Create dataframe from patches
patches_raster_df <- create_patch_df(planning_grid = planning_raster, features = features_raster, patches = seamount_patches)
# Run boundary matrix
boundary_matrix <- create_boundary_matrix(planning_grid = planning_raster, patches = seamount_patches, patch_df = patches_raster_df)
}
