% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_patch_df.R
\name{create_patch_df}
\alias{create_patch_df}
\title{Create a dataframe that counts each patch independently for the prioritizr input}
\usage{
create_patch_df(
  planning_grid,
  features,
  patches,
  costs = NULL,
  locked_out = NULL,
  locked_in = NULL
)
}
\arguments{
\item{planning_grid}{a raster or sf template with the desired resolution and coordinate reference system generated by \code{offshoredatr::get_planning_grid()}; values in areas of interest are 1, while all other values are NA}

\item{features}{a raster or sf object that includes all relevant features to be used in the prioritization; each layer of the raster or each column of the sf object identifies the location of each feature}

\item{patches}{a raster or sf object generated by \code{create_patches()}; a single feature split into spatially distinct groups; each layer of the raster or each column of the sf object identifies the location of each patch}

\item{costs}{a raster or sf object of the costs for protecting each cell in the planning grid}

\item{locked_out}{a raster or sf object for areas to be locked out (absolutely not protected) in the prioritization}

\item{locked_in}{a raster or sf object for areas to be locked in (absolutely protected) in the prioritization}
}
\value{
A data frame to be used as an input for \code{create_boundary_matrix()}
}
\description{
This function takes all of the inputs of prioritizr and creates a dataframe precursor for prioritizr. This specific function includes the use of "patches" which have been already generated by \code{create_patches()}. "Patches" are spatial groups of a particular feature, such as seamounts. Prioritizr will protect entire patches to meet the protection goal for that feature. The output of this function should be passed to \code{create_boundary_matrix()} before being passed to prioritizr.
}
\examples{
## To be updated later
# Create patches from seamount raster data
seamount_patches <- create_patches(seamount_raster)
# Create dataframe from patches
patches_df <- create_patch_df(planning_grid = planning_raster, features = features_raster, patches = seamount_patches)
}
