% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_patches.R
\name{create_patches}
\alias{create_patches}
\title{Create patches for a specific feature}
\usage{
create_patches(feature, planning_grid = NULL)
}
\arguments{
\item{feature}{a raster or sf object with the feature of interest present (value = 1) or absent (value = NA)}

\item{planning_grid}{a raster or sf template with the desired resolution and coordinate reference system generated by \code{offshoredatr::get_planning_grid()}; values in areas of interest are 1, while all other values are NA (only required if feature is a sf object)}
}
\value{
A raster or sf object with independent layers (raster)/columns (sf) designating the location of each patch
}
\description{
This function takes a feature and splits it into multiple "patches" of the feature. This is ideal for features that you want to protect x\% of, but want to make sure that entire patches (as opposed to pieces of patches) are protected to meet that target.
}
\examples{
## To be updated later
# Create patches from seamount raster data
seamount_patches <- create_patches(seamount_raster)
}
