% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_targets.R
\name{features_targets}
\alias{features_targets}
\title{Create a dataframe for the relative and absolute targets for features}
\usage{
features_targets(
  targets,
  features,
  pre_patches,
  locked_out = NULL,
  locked_in = NULL
)
}
\arguments{
\item{targets}{a vector of targets for protection (range between 0 and 1); must be the same length as the number of features + the pre-patches variable}

\item{features}{a raster or sf object that includes all relevant features to be used in the prioritization; each layer of the raster or each column of the sf object identifies the location of each feature}

\item{pre_patches}{a raster or sf object that includes the feature that is to be split into patches (not the layer that is already split into patches)}

\item{locked_out}{a raster or sf object for areas to be locked out (absolutely not protected) in the prioritization}

\item{locked_in}{a raster or sf object for areas to be locked in (absolutely protected) in the prioritization}
}
\value{
A data frame to be used to specify targets for features. Will need to be plugged into \code{constraint_targets()} before being implemented into prioritizr
}
\description{
This function creates a dataframe of targets for each of the features to be used in the prioritization
}
\examples{
## To be updated later
# Create target features - using just 20\% for every feature
features_targets <- features_targets(targets = rep(0.2, (terra::nlyr(features_raster)) + 1), features = features_raster, pre_patches = seamounts_raster)
}
